requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
seaborn==0.13.2
ydata_profiling==4.12.0
ipywidgets==8.1.5
sklearn-pandas==2.2.0
ipykernel==6.29.5
xgboost==2.1.3
fastparquet==2024.11.0
optuna==4.1.0
hyperopt==0.2.7
lightgbm==4.5.0
catboost==1.2.7
tpot==0.12.2
scipy==1.13.1
joblib==1.4.2
torch==2.5.1
deap==1.4.1
update-checker==0.18.0
tqdm==4.67.1
stopit==1.1.2
pycaret[full]==3.3.2